feat(ehr): patching signing key route
Ref: #1040
