+++ packages/api/src/routes/internal/ehr/elation/index.ts
@@ -1,7 +1,9 @@
 import Router from "express-promise-router";
 import patient from "./patient";
+import signingKey from "./signing-key";
 const routes = Router();
 
 routes.use("/patient", patient);
+routes.use("/signing-key", signingKey);
 
 export default routes;
